# [Converter]: Improve performance and add option to set maximum line length to prevent freezes