Fixes https://github.com/godotengine/godot/issues/63672

This PR added option to choose maximum checked file size and maximum length of checked line.
By default only files smaller than 4MB are checked and lines shorter than 10000 characters, which seems for me quite reasonable(at the end there is additional info how much exactly lines were ignored).

Additionally due caching regex, searching sometimes can be faster by few percent(I was expecting bigger change, but looks that most of the time is spend on searching files with regex, not creating new regex thousands of times).

When testing projects with a lot of files and short lines, I got a little lower performance than before(this was probably caused by running regexes on lines instead full files like before - this change was required to be able to cache regex), but when checking tps-demo which contains files with longer lines, time to check for renames dropped from 120s to 13s

Edit:
I found with help of hotspot that running `regex.sub` on small lines is very time consuming, so using `string.contains` before to decrease amount of regex to run can speedup this part even few times

Some benchmarks:

| Project | Mode | Master branch | PR | 
| --- | --- | --- | --- |
| tps-demo | converting | 180s | 11s |
| tps-demo | validating | 116s | 11s | 
| jetpaca | converting | 98s | 26s | 
| jetpaca | validating | 290s | 27s | 
| liblast | converting | 11s | 7s | 
| liblast | validating | 32s | 6s | 

Edit2:
Also added(re-enable actually) check to validate if new function name already exists.