@tinmanjuggernaut 2022-08-15
I tried this PR 1cabf13b1fd8bc834bc18e5e69a3b48abeef572a on a fresh copy of my gd3 project. It reports 2011 resources to do.

It is indeed much faster and no longer hangs on the file I mentioned. Great work.

First I tried with 1GB file size and 1MB on one line. It completed in a few minutes.
```
../../godot/bin/godot.windows.opt.tools.64.exe --convert-3to4 999999 999999
...
Conversion ended - all files(2011), converted files(2011), not converted files(0).
Conversion of all files took 462119 ms.
```

However, I had many instances of skipped lines. Yet none of them are displayed at the end, neither in summary nor detail. Those are now broken files. It should report the skipped lines files at the end so a user can take steps to fix them.

It even processed my 353MB player scene file with 200 animations, a complex rig, and complete AAA quality human model. 88 seconds to process it is acceptable given what it is.
```
Trying to convert       1714/2011 file - "assets/models/characters/dorian/Dorian.tscn" with size - 353159 KB
    Checking file took 88459 ms.
```

If it can process the above file, then it should be able to process any other file I have. The default 10,000 character limit is wayyyyy too low. 999,999 was fine.

Yet it skipped lines > 1MB as I specified. 
```
Trying to convert       329/2011 file - "src/inventory/items/shields/OrnateBuckler.tscn" with size - 1429 KB
    Checking file took 219 ms.
    File was changed, conversion took 219 ms.
    Ignored 1 lines, because their length exceeds maximum allowed characters - 999999
```

So I expanded it to 100MB which it should be able to allocate and process without sweating. However it hung on a tiny 4mb file. The whole thing could fit in an allocation smaller than many texture files. It's the same structure as described in #63672 with a LOD manger, a few ArrayMeshes.

```
../../godot/bin/godot.windows.opt.tools.64.exe --convert-3to4 999999 99999999

Trying to convert       361/2010 file - "scenes/BlackMountain.tscn" with size - 4354 KB

# after Ctrl+C
ERROR: Pages in use exist at exit in PagedAllocator: N7Variant5Pools11BucketLargeE
   at: (./core/templates/paged_allocator.h:140)
ERROR: Pages in use exist at exit in PagedAllocator: N7Variant5Pools12BucketMediumE
   at: (./core/templates/paged_allocator.h:140)
ERROR: Pages in use exist at exit in PagedAllocator: N7Variant5Pools11BucketSmallE
   at: (./core/templates/paged_allocator.h:140)
ERROR: BUG: Unreferenced static string to 0: interface_added
   at: unref (core/string/string_name.cpp:131)
ERROR: BUG: Unreferenced static string to 0: bus_layout_changed
   at: unref (core/string/string_name.cpp:131)

```

I left this for about 10-15 minutes. Allocated memory is only 185mb. Processor is 100% on one core. 350mb Dorian only took 88 seconds and the process allocated up to 3GB.