@qarmin 2022-08-17
Latest freeze with 4MB file should be fixed now.

After recent changes, String `contains` method takes ~75% converting time so I don't think that there is room for improvements
The only two possible things which could be done to speedup a little a code:
- caching rest of regex, this will increase slightly startup time
- improving algorithm of finding text in strings(https://en.wikipedia.org/wiki/String-searching_algorithm#Single-pattern_algorithms)

![Screenshot from 2022-08-17 20-55-00](https://user-images.githubusercontent.com/41945903/185221999-965b7936-c6eb-4560-8fe4-e650d0d52afa.png)