@tinmanjuggernaut 2022-08-18
Ok, I let it work again on my project w/ 1GB files and 100MB lines:

```
../../godot/bin/godot.windows.opt.tools.64.exe --convert-3to4 999999 99999999

Trying to convert       570/2010 file - "assets/models/village/buildings/HouseC.tscn" with size - 60322 KB
    Checking file took 30839 ms.
    File was changed, conversion took 30839 ms.

Trying to convert       1614/2010 file - "assets/models/dungeon/architecture/TrippleArch.tscn" with size - 15323 KB
    Checking file took 8674 ms.
    File was changed, conversion took 8674 ms.

Trying to convert       1679/2010 file - "assets/models/creatures/dragon/dragon_body.tres" with size - 13157 KB
    Checking file took 1408 ms.
    File was changed, conversion took 1408 ms.

Trying to convert       1713/2010 file - "assets/models/characters/dorian/Dorian.tscn" with size - 353159 KB
    Checking file took 184101 ms.
    File was changed, conversion took 184101 ms.

Conversion ended - all files(2010), converted files(2010), not converted files(0).
Conversion of all files took 558289 ms.

```

Nothing was skipped, nothing hung. It chew through even monster files like the highlights included above. Dorian took twice as long, but still acceptable. Great work. Thank you!