@qarmin 2022-08-18
Recently I only added changes which should speedup searching, especially with big files, so 2x slower checking of 300mb file is quite suspicious.

I think that converting times cannot be really predictable due random OS CPU usage(converter should most of the time use 100% of one core), because one time entire godot demo repository converted most of the time in 32s but sometimes took even in 47s.

The only possible way now to speedup searching is to move computations to multiple cores(which may I implement in future in other PR)