@tinmanjuggernaut 2022-08-18
Converting my 4gb project in 10 minutes is fine. The skipping and hanging issues were a bigger deal and those seem resolved.